+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Improve the `exclude` setting to allow filenames, extensions, and patterns <https://github.com/gtronset/beets-filetote/pull/179>
 
+### Fixed
+
+- Bugfix: allow list of strings for deprecated `exclude` config <https://github.com/gtronset/beets-filetote/pull/180>
+
 ## [1.0.1] - 2025-05-08
 
 ### Changed